backend/app/core/db.py
from functools import lru_cache

from sqlalchemy.ext.asyncio import AsyncEngine, create_async_engine
from sqlmodel import select
from sqlmodel.ext.asyncio.session import AsyncSession

from app import crud
from app.core.config import settings
from app.models import User, UserCreate


@lru_cache(maxsize=1)
def get_engine() -> AsyncEngine:
    """Get or create the database engine (lazy singleton)."""
    return create_async_engine(str(settings.SQLALCHEMY_DATABASE_URI), echo=False)


# make sure all SQLModel models are imported (app.models) before initializing DB
# otherwise, SQLModel might fail to initialize relationships properly
# for more details: https://github.com/fastapi/full-stack-fastapi-template/issues/28


async def init_db(session: AsyncSession) -> None:
    # Tables should be created with Alembic migrations
    # Models are already imported and registered from app.models
    result = await session.exec(select(User).where(User.email == settings.FIRST_SUPERUSER))
    user = result.first()
    if not user:
        user_in = UserCreate(
            email=settings.FIRST_SUPERUSER,
            password=settings.FIRST_SUPERUSER_PASSWORD,
            is_superuser=True,
        )
        user = await crud.create_user(session=session, user_create=user_in)
